xu response cho dep

+++ src/main/java/Nhom02/Nhom02HappyFarm/api/blog/BlogApi.java
@@ -2,18 +2,16 @@ package Nhom02.Nhom02HappyFarm.api.blog;
 
 
 import Nhom02.Nhom02HappyFarm.entities.Blog;
+import Nhom02.Nhom02HappyFarm.response.ResponseHandler;
 import Nhom02.Nhom02HappyFarm.service.BlogService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.server.ResponseStatusException;
-
-import java.util.List;
+import java.util.Optional;
 
 @RequiredArgsConstructor
 @RestController
@@ -21,18 +19,18 @@ import java.util.List;
 @Api(value = "Api blog")
 public class BlogApi {
     private final BlogService blogService;
-
+    private final ResponseHandler responseHandler;
     @ApiOperation(value = "Lay 1 list cac blog")
     @ApiResponses(value = {
             @ApiResponse(code = 200, message = "Lay thanh cong"),
             @ApiResponse(code = 400, message = "Co loi xay ra trong qua trinh lay")
     })
     @GetMapping("/getlist")
-    public ResponseEntity<List<Blog>> getList(){
+    public ResponseEntity<Object> getList(){
         try {
-            return new ResponseEntity<>(blogService.getAllBlog(), HttpStatus.OK);
+            return ResponseEntity.ok(responseHandler.successResponse("Get list success", blogService.getAllBlog()));
         }catch (Exception e){
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
+            return ResponseEntity.badRequest().body(responseHandler.failResponse(e.getMessage()));
         }
     }
     @ApiOperation(value = "Tao 1 blog moi de them ")
@@ -41,25 +39,37 @@ public class BlogApi {
             @ApiResponse(code = 400, message = "Co loi xay ra trong qua trinh lay")
     })
     @GetMapping("/addblog")
-    public ResponseEntity<Blog> addNewBlog(){
+    public ResponseEntity<Object> addNewBlog(){
         try {
-            return new ResponseEntity<>(new Blog(), HttpStatus.OK);
+            return ResponseEntity.ok(responseHandler.successResponse("Create success", new Blog()));
         }catch (Exception e){
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
+            return ResponseEntity.badRequest().body(responseHandler.failResponse(e.getMessage()));
         }
     }
-    @ApiOperation(value = "Them moi 1 blog")
+    @ApiOperation(value = "Them moi 1 blog, request body la 1 blog")
     @ApiResponses(value = {
             @ApiResponse(code = 200, message = "Thanh cong"),
             @ApiResponse(code = 400, message = "Co loi xay ra trong qua trinh lay")
     })
     @PostMapping("/addblog")
-    public ResponseEntity<Blog> addNewBlog(@RequestBody Blog blog){
+    public ResponseEntity<Object> addNewBlog(@RequestBody Blog blog){
         try {
+            if(blog.getDetails().isEmpty()){
+                return ResponseEntity.badRequest().body(responseHandler.failResponse("Vui long nhap chi tiet blog"));
+            }else if (blog.getTitle().isEmpty()){
+                return ResponseEntity.badRequest().body(responseHandler.failResponse("Vui long nhap tieu de"));
+            }else if(blog.getUserCreate() == null) {
+                return ResponseEntity.badRequest().body(responseHandler.failResponse("Khong tim thay idUser"));
+            }else {
+                Optional<Blog> check = blogService.getAllBlog().stream().filter(s -> s.getTitle().equals(blog.getTitle())).findFirst();
+                if (check.isPresent()){
+                    return ResponseEntity.badRequest().body(responseHandler.failResponse("Tieu de bai viet bi trung"));
+                }
+            }
             blogService.createOrSaveBlog(blog);
-            return new ResponseEntity<>(HttpStatus.OK);
+            return ResponseEntity.ok(responseHandler.successResponse("Tao moi thanh cong", blog));
         }catch (Exception e){
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
+            return ResponseEntity.badRequest().body(responseHandler.failResponse(e.getMessage()));
         }
     }
     @ApiOperation(value = "Lay blog theo {id}")
@@ -68,11 +78,14 @@ public class BlogApi {
             @ApiResponse(code = 400, message = "Co loi xay ra trong qua trinh lay")
     })
     @GetMapping("/getblog/{id}")
-    public ResponseEntity<Blog> getBlogById(@PathVariable String id){
+    public ResponseEntity<Object> getBlogById(@PathVariable String id){
         try{
-            return new ResponseEntity<>(blogService.getBlog(id), HttpStatus.OK);
+            if (blogService.getBlog(id) == null){
+                return ResponseEntity.badRequest().body(responseHandler.failResponse("Not found"));
+            }
+            return ResponseEntity.ok(responseHandler.successResponse("Tim blog thanh cong", blogService.getBlog(id)));
         }catch (Exception e){
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
+            return ResponseEntity.badRequest().body(responseHandler.failResponse(e.getMessage()));
         }
     }
     @ApiOperation(value = "Edit blog theo {id}")
@@ -80,13 +93,17 @@ public class BlogApi {
             @ApiResponse(code = 200, message = "Thanh cong"),
             @ApiResponse(code = 400, message = "Co loi xay ra trong qua trinh lay")
     })
-    @PutMapping("/edit/{id}")
-    public ResponseEntity<Blog> editBlog(@PathVariable String id, @RequestBody Blog blog){
+    @PutMapping("/edit")
+    public ResponseEntity<Object> editBlog(@RequestBody Blog blog){
         try{
+            Blog check = blogService.getBlog(blog.getIdBlog());
+            if (check == null){
+                return ResponseEntity.badRequest().body(responseHandler.failResponse("Not found"));
+            }
             blogService.createOrSaveBlog(blog);
-            return new ResponseEntity<>(HttpStatus.OK);
+            return ResponseEntity.ok(responseHandler.successResponse("Edit thanh cong", blog));
         }catch (Exception e){
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
+            return ResponseEntity.badRequest().body(responseHandler.failResponse(e.getMessage()));
         }
     }
     @ApiOperation(value = "Xoa blog theo {id}")
@@ -95,12 +112,15 @@ public class BlogApi {
             @ApiResponse(code = 400, message = "Co loi xay ra trong qua trinh lay")
     })
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<Blog> deleteBlog(@PathVariable String id){
+    public ResponseEntity<Object> deleteBlog(@PathVariable String id){
         try{
+            if(blogService.getBlog(id) == null){
+                return ResponseEntity.badRequest().body(responseHandler.failResponse("Not found"));
+            }
             blogService.deleteBlog(id);
-            return new ResponseEntity<>(HttpStatus.OK);
+            return ResponseEntity.ok(responseHandler.successResponse("Xoa thanh cong", id));
         }catch (Exception e){
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
+            return ResponseEntity.badRequest().body(responseHandler.failResponse(e.getMessage()));
         }
     }
 }
