+++ src/main/java/Nhom02/Nhom02HappyFarm/api/brand/BrandApi.java
@@ -2,39 +2,41 @@ package Nhom02.Nhom02HappyFarm.api.brand;
 
 
 import Nhom02.Nhom02HappyFarm.entities.Brand;
-import Nhom02.Nhom02HappyFarm.entities.OriginFertilizer;
+import Nhom02.Nhom02HappyFarm.response.ResponseHandler;
 import Nhom02.Nhom02HappyFarm.service.BrandService;
-import Nhom02.Nhom02HappyFarm.service.OriginService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.server.ResponseStatusException;
-
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/api/brand")
 @RequiredArgsConstructor
 @CrossOrigin(origins = "http://localhost:8080")
 @Api(value = "Manage Brand Of Fertilizer")
+
 public class BrandApi {
     private final BrandService brandService;
+    private final ResponseHandler responseHandler;
     @ApiOperation(value = "Tra ve 1 list cac thuong hieu bao gom ca xoa va chua xoa")
     @ApiResponses(value ={
         @ApiResponse(code = 200, message = "Tra ve thanh cong"),
         @ApiResponse(code = 204, message = "List rong"), @ApiResponse(code = 400, message = "Lỗi khi lấy list")
     })
     @GetMapping("/getlistbrand")
-    public ResponseEntity<List<Brand>> getListBrand(@RequestParam(required = false) String nameBrand){
+    public ResponseEntity<Object> getListBrand(@RequestParam(required = false) String nameBrand){
         if (brandService.GetAllBrand(nameBrand).isEmpty()){
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            return ResponseEntity.ok(responseHandler.successResponseButNotHaveContent("List brand rong"));
         }
-        return new ResponseEntity<>(brandService.GetAllBrand(nameBrand), HttpStatus.OK);
+        return ResponseEntity.ok(responseHandler.successResponse("Lay list thanh cong", brandService.GetAllBrand(nameBrand)));
     }
     @ApiOperation(value = "Thêm mới 1 brand vào db")
     @ApiResponses(value ={
@@ -42,12 +44,15 @@ public class BrandApi {
             @ApiResponse(code = 401, message = "Có lỗi ve du lieu trong quá trình thêm, kiểm tra xem có dữ liệu nào null hay không")
     })
     @PostMapping("/addnew")
-    public ResponseEntity<Brand> createNewBrand(@RequestBody Brand brand){
+    public ResponseEntity<Object> createNewBrand(@RequestBody Brand brand){
         try{
+            if (brand.getNameBrand().isEmpty()){
+                return ResponseEntity.badRequest().body(responseHandler.failResponse("Not found name brand"));
+            }
             brandService.AddOrEditBrand(brand);
-            return new ResponseEntity<>(HttpStatus.CREATED);
+            return ResponseEntity.ok(responseHandler.successResponse("Create succesfully", brand));
         }catch (Exception error) {
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            return ResponseEntity.badRequest().body(responseHandler.failResponse(error.getMessage()));
         }
     }
     @ApiOperation(value = "Tìm 1 brand với tên trong DB (ten phải nhập đúng thì mới trả ra kết quả)")
@@ -56,12 +61,12 @@ public class BrandApi {
             @ApiResponse(code = 401, message = "Có lỗi ve du lieu trong quá trình thêm, kiểm tra xem có dữ liệu nào null hay không")
     })
     @GetMapping("/find/{name}")
-    public ResponseEntity<List<Brand>> findByName(@PathVariable(name = "name") String name){
+    public ResponseEntity<Object> findByName(@PathVariable(name = "name") String name){
         List<Brand> list = brandService.findByName(name);
-        if(list.isEmpty()){
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        if(list.isEmpty()) {
+            return ResponseEntity.ok(responseHandler.successResponseButNotHaveContent("List rong"));
         }
-        return new ResponseEntity<>(list, HttpStatus.OK);
+        return ResponseEntity.ok(responseHandler.successResponse("Succesfully get list", list));
     }
 
     @ApiOperation(value = "Lay 1 brand")
@@ -70,11 +75,11 @@ public class BrandApi {
             @ApiResponse(code = 400, message = "Co loi xay ra trong qua trinh lay du lieu")
     })
     @GetMapping("/newBrand")
-    public ResponseEntity<Brand> createNew(){
+    public ResponseEntity<Object> createNew(){
         try{
-            return new ResponseEntity<>(new Brand(), HttpStatus.OK);
+            return ResponseEntity.ok(responseHandler.successResponse("Create new Brand", new Brand()));
         }catch(Exception e){
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
+            return ResponseEntity.badRequest().body(responseHandler.failResponse(e.getMessage()));
         }
     }
 
@@ -84,29 +89,34 @@ public class BrandApi {
             @ApiResponse(code = 400, message = "Co loi xay ra trong qua trinh lay du lieu")
     })
     @GetMapping("/getbrand/{id}")
-    public ResponseEntity<Brand> getBrand(@PathVariable String id){
+    public ResponseEntity<Object> getBrand(@PathVariable String id){
         try{
             Brand brand = brandService.GetBrand(id);
-            return new ResponseEntity<>(brand, HttpStatus.OK);
+            if (brand == null){
+                return ResponseEntity.badRequest().body(responseHandler.failResponse("Not found brand"));
+            }
+            return ResponseEntity.ok(responseHandler.successResponse("Get Brand successfully", brand));
         }catch(Exception e){
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
+            return ResponseEntity.badRequest().body(responseHandler.failResponse(e.getMessage()));
         }
     }
-
     @ApiOperation(value = "Chỉnh sua brand trong DB")
     @ApiResponses(value ={
             @ApiResponse(code = 200, message = "Chinh sua thanh cong"),
             @ApiResponse(code = 404, message = "Không tìm thay Id cua brand cần chỉnh sửa")
     })
     @PutMapping("/editbrand/{id}")
-
-    public ResponseEntity<Brand> editBrand(@PathVariable(name = "id") String idBrand, @RequestBody Brand brand){
+    public ResponseEntity<Object> editBrand(@PathVariable(name = "id") String idBrand, @RequestBody Brand brand){
         Brand getBrand = brandService.GetBrand(idBrand);
-        if (getBrand == null){
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }else {
-            brandService.AddOrEditBrand(brand);
-            return new ResponseEntity<>(HttpStatus.OK);
+        try {
+            if (getBrand == null) {
+                return ResponseEntity.badRequest().body(responseHandler.failResponse("Not found brand"));
+            } else {
+                brandService.AddOrEditBrand(brand);
+                return ResponseEntity.ok(responseHandler.successResponse("Edit brand success", brand));
+            }
+        }catch (Exception e){
+            return ResponseEntity.badRequest().body(responseHandler.failResponse(e.getMessage()));
         }
     }
     @ApiOperation(value = "Xóa 1 brand với ID là tham số truyền vào")
@@ -115,12 +125,16 @@ public class BrandApi {
             @ApiResponse(code = 400, message = "Không tìm thay Id cua brand cần xoa")
     })
     @DeleteMapping("/deletebrand/{id}")
-    public ResponseEntity<Brand> deteleBrand(@PathVariable(name = "id") String idBrand){
+    public ResponseEntity<Object> deteleBrand(@PathVariable(name = "id") String idBrand){
         try {
+            Brand brand = brandService.GetBrand(idBrand);
+            if(brand == null){
+                return ResponseEntity.badRequest().body(responseHandler.failResponse("Not found brand"));
+            }
             brandService.DeleteBrand(idBrand);
-            return new ResponseEntity<>(HttpStatus.OK);
+            return ResponseEntity.ok(responseHandler.successResponse("Delete success", brand));
         }catch (Exception error){
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            return ResponseEntity.badRequest().body(responseHandler.failResponse(error.getMessage()));
         }
     }
     @ApiOperation(value = "Trả về các brand với IsDelete = false (Trả ve cac brand chưa xóa)")
@@ -130,11 +144,11 @@ public class BrandApi {
             @ApiResponse(code = 204, message = "List rỗng"),
     })
     @GetMapping("/notdeletebrand")
-    public ResponseEntity<List<Brand>> brandIsNotDelete(){
+    public ResponseEntity<Object> brandIsNotDelete(){
         List<Brand> listBrand = brandService.GetAllOriginNotDelete();
         if (listBrand.isEmpty()){
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            return ResponseEntity.ok(responseHandler.successResponseButNotHaveContent("List rong"));
         }
-        return new ResponseEntity<>(listBrand, HttpStatus.OK);
+        return ResponseEntity.ok(responseHandler.successResponse("Success get list", listBrand));
     }
 }
